+++ dialect/sql/schema/postgres.go
@@ -75,7 +75,7 @@ func (d *Postgres) fkExist(ctx context.Context, tx dialect.Tx, name string) (boo
 }
 
 // setRange sets restart the identity column to the given offset. Used by the universal-id option.
-func (d *Postgres) setRange(ctx context.Context, tx dialect.Tx, t *Table, value int) error {
+func (d *Postgres) setRange(ctx context.Context, tx dialect.Tx, t *Table, value int64) error {
 	if value == 0 {
 		value = 1 // RESTART value cannot be < 1.
 	}
